add auto count interval

+++ core/src/main/java/com/soapgu/core/ICounter.java
@@ -8,5 +8,5 @@ public interface ICounter {
      * 获取当前计数
      * @return 计数值
      */
-    int getCount();
+    Long getCount();
 }
